Draft: move more functions to draftgeoutils.intersections
